+++ contest/ABC405/abc405d.py
@@ -1,3 +1,6 @@
+# ABC405D Escape Route
+# 関数定義スペース
+
 class Board():
     """
     二次元ボードを便利に使いやすくするためのクラス
@@ -210,3 +213,63 @@ def input_board_with_wall(height:int,weight:int,wall:any,f:lambda x:input(list()
         board.append(f()+[wall])
     board.append([wall]*(weight+1))
     return Board(board)
+
+
+
+
+
+def main():
+    # 入力スペース ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~Lit_to
+    H, W = map(int,input().split())
+    BOARD=input_board_with_wall(H,W,"#",lambda:list(input()))
+    # 処理スペース ================================================================================================Lit_to
+    board=BOARD.copy()
+
+    queue=deque()
+    for i in range(H):
+        for j in range(W):
+            if BOARD[(i,j)]=="E":
+                queue.append((i,j))
+
+    while queue:
+        pos=queue.popleft()
+        for j in LRUD:
+            new_pos=(pos[0]+j[0],pos[1]+j[1])
+            new_data=board.look(new_pos)
+            if new_data == ".":
+                queue.append(new_pos)
+                board[new_pos]=j[2]
+
+    for i in range(H):
+        for j in range(W):
+            print(board[(i,j)],end="")
+        print()
+
+
+# 以下コピペ部分=============================================================
+# インポート(本編はだいたい30行目あたり)
+import sys,itertools,math,heapq
+from collections import defaultdict,deque
+from sortedcontainers import SortedSet, SortedList, SortedDict # CPython?
+# pypyjit.set_param('max_unroll_recursion=-1')
+sys.setrecursionlimit(10**8)
+sys.set_int_max_str_digits(0)
+dict=defaultdict
+# 便利定数定義
+ALPHABET="ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+# ALPHABET="abcdefghijklmnopqrstuvwxyz"
+MOD=998244353
+MAX=10**18
+LRUD=[(0,-1,">"),(0,1,"<"),(-1,0,"v"),(1,0,"^")]
+LURULDRD=[(-1,-1),(-1,1),(1,-1),(1,1)]
+# 便利関数定義
+def input(): return (sys.stdin.readline()).rstrip()
+def printe(*values,sep=" ",end="\n"):print(*values,sep=sep,end=end);exit()
+def yes(f=True): printe("Yes") if (f) else None
+def no(f=True): printe("No") if (f) else None
+def listr(l:list,f=str): return "".join(list(map(f,l)))
+def debug(*values,sep=" ",end="\n"): print(*values,sep=sep,end=end,file=sys.stderr)
+def look(pos:tuple,board:list): return board[pos[0]][pos[1]]
+def printYN(f:bool): yes() if f else no()
+if __name__=="__main__":
+    main()
